clean: Improvements based on CodeQL analysis
Cleaning up some code based on CodeQL analysis results. These are all related to comparing a smaller integer type with a wider type. This should prevent possible infinite loops, although these were likely not a problem to begin with due to how all of these things are reported by devices and lengths in the spec being far smaller than something to worry about, but this will protect from weird cases where things are not reporting properly.
Lots of C_CAST to make sure VS doesn't produce lots of warnings.

+++ utils/C/openSeaChest/openSeaChest_Info.c
@@ -187,7 +187,7 @@ int32_t main(int argc, char *argv[])
             }
             else if (strncmp(longopts[optionIndex].name, SCSI_DEFECTS_LONG_OPT_STRING, M_Min(strlen(longopts[optionIndex].name), strlen(SCSI_DEFECTS_LONG_OPT_STRING))) == 0)
             {
-                uint8_t counter = 0;
+                size_t counter = 0;
                 SCSI_DEFECTS_FLAG = true;
                 while (counter < strlen(optarg))
                 {
@@ -638,7 +638,7 @@ int32_t main(int argc, char *argv[])
     else
     {
         /*need to go through the handle list and attempt to open each handle.*/
-        for (uint16_t handleIter = 0; handleIter < DEVICE_LIST_COUNT; ++handleIter)
+        for (uint32_t handleIter = 0; handleIter < DEVICE_LIST_COUNT; ++handleIter)
         {
             /*Initializing is necessary*/
             deviceList[handleIter].sanity.size = sizeof(tDevice);
